requirements.txt
# Core dependencies
playwright==1.40.0
pandas==2.1.4
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Async support
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Data validation and processing
validators==0.22.0
tldextract==5.1.1

# Configuration and logging
pyyaml==6.0.1
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional: For more advanced email detection
regex==2023.10.3 